Create READ.md
This is an analysis of what viewers think about “House of the Dragon” series which is a spinoff of “Game of thrones” (GoT), arguably the biggest tv show ever created. 
The show is set some 172 years before the birth of Daenerys Targaryen who is one of the main characters of GoT. It depicts the reign of house Targaryen and what ultimately led to the collapse of their all-powerful house. 
The show was premiered on August 21, 2022. 
The analysis seeks to reveal insights like most talked about cast, most predominant hashtags and others.